requirements.txt
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
alembic==1.13.1
aiogram==3.5.0
